Moved and splitted bellcurve series into a subfolder.
