+++ ts/Series/Bellcurve/BellcurvePoint.d.ts
@@ -1,0 +1,40 @@
+/* *
+ *
+ *  (c) 2010-2020 Highsoft AS
+ *
+ *  Author: Sebastian Domas
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type AreaSplinePoint from '../AreaSpline/AreaSplinePoint';
+import type BellcurvePointOptions from './BellcurvePointOptions';
+import type BellcurveSeries from './BellcurveSeries';
+
+/* *
+ *
+ *  Class
+ *
+ * */
+
+export class BellcurvePoint extends AreaSplinePoint {
+    public option: BellcurvePointOptions;
+    public series: BellcurveSeries;
+}
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
+
+export default BellcurvePoint;
