+++ ts/Series/Bellcurve/BellcurvePointOptions.d.ts
@@ -1,0 +1,31 @@
+/* *
+ *
+ *  (c) 2010-2020 Highsoft AS
+ *
+ *  Author: Sebastian Domas
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type AreaSplinePointOptions from '../AreaSpline/AreaSplinePointOptions';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface BellcurvePointOptions extends AreaSplinePointOptions {
+    // nothing here yet
+}
+
+export default BellcurvePointOptions;
